# api/openapi.yaml
openapi: 3.1.0
info:
  title: Snooze API
  version: 2.0.0
  summary: Log aggregation and alerting REST API.
  description: |
    The Snooze v1 HTTP API. Snooze 2.0 is a Go rewrite of the Python 1.x
    server; this document describes the **v1** wire contract that ships
    with 2.0.

    ## Authentication

    Every endpoint outside the small "public" set (`/healthz`,
    `/readyz`, `/metrics`, `/api/v1/login/*`) requires a bearer JWT:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `POST /api/v1/login/{local,ldap,anonymous}`.
    The Python 1.x `Authorization: JWT <token>` scheme is no longer
    accepted.

    ## Multi-tenancy

    Snooze supports multiple isolated organizations (tenants) on a single
    server. Every authenticated request is scoped to the tenant encoded in
    the `tenant_id` JWT claim. Pass the `org` field in the login request
    body to obtain a token scoped to a specific tenant; omit `org` (or leave
    it blank) to land in the `default` tenant.

    **Ingest routing** — unauthenticated alert ingestion (`POST /api/v1/alerts`,
    `POST /api/v1/webhook/*`) is routed to a tenant by the `Authorization:
    Bearer <ingest_token>` header, where `<ingest_token>` is the per-tenant
    opaque token stored in the tenant document. When no token is supplied or
    the token is unknown, the request falls back to the `default` tenant.

    **Platform permissions** — the `ro_tenant` and `rw_tenant` permissions gate
...

    ## CRUD shape
...
servers:
...
security:
  - bearerAuth: []
  # Trusted-header (auth-proxy) mode is an alternative to a Bearer token on
  # every authenticated route, but only when auth_proxy.enabled=true and the
  # immediate TCP peer is a trusted proxy. Disabled by default.
  - proxyHeader: []
...
paths:
...
  /api/v1/user/me/password:
    post:
      tags: [auth]
      summary: Self-service password change
      description: |
        Replaces the authenticated user's password. Only available when
        the caller's JWT `method` claim is `local` — LDAP / anonymous
        accounts are managed by their backing provider.

        The caller MUST supply their current password; the server
        re-runs the bcrypt comparison through the same code path as
        `/api/v1/login/local`, so the failure mode (and timing) is
        identical to a bad-credentials login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password]
              properties:
                current_password:
                  type: string
                  description: The caller's current password.
                password:
                  type: string
                  description: The new password (bcrypt-hashed server-side).
      responses:
        '204':
          description: Password updated.
        '401':
          description: Missing claims or wrong current password.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
        '403':
          description: |
            Caller's auth method is not `local`; self-service password
            change is not available for that backend.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
        '422':
          description: Empty body field(s).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
  /api/v1/user/me/apikeys:
    get:
      tags: [auth]
      summary: List my API keys
      responses:
        '200':
          description: The caller's API keys (secrets stripped).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/APIKey'}
    post:
      tags: [auth]
      summary: Create an API key for myself
      description: >
        Mints a key scoped to the caller, carrying a subset of the caller's own
        permissions. The raw key is returned once. Not permitted when the
        request is itself authenticated with an API key.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/APIKeyCreate'}
      responses:
        '201':
          description: Created. Body includes the raw `key` (shown once).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/APIKeyCreated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422':
          description: Invalid expiry, or requested permissions are malformed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
  /api/v1/user/me/apikeys/{id}:
    delete:
      tags: [auth]
      summary: Revoke one of my API keys
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        '204':
          description: Deleted.
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/user/me/avatar:
    put:
      tags: [auth]
      summary: Upload my profile picture
      description: |
        Stores the caller's profile picture. The body carries a `data:` URL
        (`data:image/png;base64,…` or `data:image/jpeg;base64,…`). The server
        decodes it, rejects oversized or malformed input, and re-encodes it to
        PNG (dropping any metadata) before storing. SVG is never accepted.
        Limits: 512 KiB decoded input, at most 512×512 pixels, aspect ratio at
        most 2:1, 128 KiB after re-encoding. Not permitted when the request is
        authenticated with an API key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: string
                  description: A `data:image/png` or `data:image/jpeg` base64 URL.
      responses:
        '200':
          description: Stored.
          content:
            application/json:
              schema:
                type: object
                required: [version]
                properties:
                  version:
                    type: string
                    description: Content hash; changes whenever the picture changes.
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413':
          description: Picture too large.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
        '422':
          description: Not a decodable PNG/JPEG data URL, or dimensions out of bounds.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
    delete:
      tags: [auth]
      summary: Remove my profile picture
      responses:
        '204':
          description: Removed (also returned when there was none).
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403':
          description: The request is authenticated with an API key.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. The Python 1.x `Authorization: JWT <token>` scheme
        is rejected. Obtain a token via `POST /api/v1/login/local`.

    ingestToken:
      type: http
      scheme: bearer
      description: |
        Per-tenant ingest token. Opaque random string stored in the tenant
        document (`ingest_token` field). Supply as `Authorization: Bearer
        <ingest_token>` or `?token=<ingest_token>`. When absent or unknown,
        the middleware falls back to the `default` tenant (D4). This scheme
        gates `/api/v1/alerts` and `/api/v1/webhook/*` in multi-tenant
        deployments; single-tenant deployments leave ingest open (existing
        behaviour preserved).

    apiKeyAuth:
      type: http
      scheme: bearer
      description: >
        A user API key, presented as `Authorization: Bearer snz_…`. Keys are
        minted at `/api/v1/user/me/apikeys`, carry a subset of the owner's
        permissions, and are bounded live by the owner's current permissions on
        every request.

    proxyHeader:
      type: apiKey
      in: header
      name: X-Forwarded-User
      description: >
        Trusted-header (auth-proxy) mode. Gated by `auth_proxy.enabled` and
        **disabled by default**. When enabled, an upstream reverse proxy
        (oauth2-proxy / Pomerium / mod_auth_openidc) that has already
        authenticated the user forwards the identity in the configured username
        header (default `X-Forwarded-User`) and an optional groups header
        (default `X-Forwarded-Groups`). Snooze JIT-provisions a `method=proxy`
...
  responses:
    BadRequest:
      description: Malformed request body or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
    Forbidden:
      description: Authenticated but lacking the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
    Conflict:
      description: Conflict (e.g. attempt to delete a reserved resource).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
...
  schemas:
...
    APIKey:
      type: object
      properties:
        uid: {type: string}
        owner: {type: string}
        owner_method: {type: string}
        name: {type: string}
        key_prefix:
          type: string
          description: First 12 chars of the key, for identification.
        permissions:
          type: array
          items: {type: string}
        groups:
          type: array
          items: {type: string}
        created_at: {type: number}
        expires_at: {type: number}
        revoked_at: {type: number}
        last_used_at:
          type: number
          description: >
            Unix epoch (seconds) of the last successful authentication with
            this key. Absent (or 0) if the key has never been used.
        use_count:
          type: number
          description: >
            Total number of times this key has authenticated a request
            (throttled: only incremented once per hour, so the value is a
            lower-bound estimate).
    APIKeyCreate:
      type: object
      required: [name]
      properties:
        name: {type: string}
        permissions:
          type: array
          items: {type: string}
          description: Must be a subset of the caller's own permissions.
        expires_at:
          type: string
          format: date-time
          description: RFC3339. Omit to default to the configured maximum lifetime.
    APIKeyCreated:
      allOf:
        - $ref: '#/components/schemas/APIKey'
        - type: object
          properties:
            key:
              type: string
              description: The raw key (snz_…). Returned ONCE, on creation only.
...
    ErrEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrBody'
    ErrBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - validation_error
            - unavailable
            - internal
        message: {type: string}
        details:
          type: object
          additionalProperties: true
        request_id: {type: string}
...